Removes all references of System.Web.Security from core project
