+++ src/Umbraco.Abstractions/DictionaryExtensions.cs
@@ -4,8 +4,8 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Linq;
+using System.Net;
 using System.Text;
-using System.Web;
 
 namespace Umbraco.Core
 {
@@ -251,7 +251,7 @@ namespace Umbraco.Core
             var builder = new StringBuilder();
             foreach (var i in d)
             {
-                builder.Append(String.Format("{0}={1}&", HttpUtility.UrlEncode(i.Key), i.Value == null ? string.Empty : HttpUtility.UrlEncode(i.Value.ToString())));
+                builder.Append(String.Format("{0}={1}&", WebUtility.UrlEncode(i.Key), i.Value == null ? string.Empty : WebUtility.UrlEncode(i.Value.ToString())));
             }
             return builder.ToString().TrimEnd('&');
         }
