+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -371,9 +371,6 @@
     <Compile Include="Models\MediaExtensions.cs" />
     <Compile Include="Models\MediaType.cs" />
     <Compile Include="Models\Member.cs" />
-    <Compile Include="Models\Membership\MembershipUserExtensions.cs" />
-    <Compile Include="Models\Membership\UmbracoMembershipMember.cs" />
-    <Compile Include="Models\Membership\UmbracoMembershipUser.cs" />
     <Compile Include="Models\Membership\User.cs" />
     <Compile Include="Models\MemberType.cs" />
     <Compile Include="Models\PublishedContent\PublishedContentTypeFactory.cs" />
@@ -745,11 +742,9 @@
     <Compile Include="Security\IUserAwarePasswordHasher.cs" />
     <Compile Include="Security\IUserSessionStore.cs" />
     <Compile Include="Security\MachineKeyGenerator.cs" />
-    <Compile Include="Security\MembershipProviderBase.cs" />
     <Compile Include="Security\ConfiguredPasswordValidator.cs" />
     <Compile Include="Security\UmbracoBackOfficeIdentity.cs" />
     <Compile Include="Security\UmbracoEmailMessage.cs" />
-    <Compile Include="Security\UmbracoMembershipProviderBase.cs" />
     <Compile Include="Security\UserAwarePasswordHasher.cs" />
     <Compile Include="Serialization\CaseInsensitiveDictionaryConverter.cs" />
     <Compile Include="Serialization\ForceInt32Converter.cs" />
