+++ src/Umbraco.Web/Models/Membership/UmbracoMembershipMember.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Web.Security;
-using Umbraco.Core.Configuration;
+using Umbraco.Core.Models.Membership;
 
-namespace Umbraco.Core.Models.Membership
+namespace Umbraco.Web.Models.Membership
 {
 
     internal class UmbracoMembershipMember : MembershipUser
